+++ moderatorFrontend/src/jsMain/kotlin/views/accessManagement/accessManagementDetails/AccessManagementDetailsController.kt
@@ -1,4 +1,4 @@
-package views.accessManagement
+package views.accessManagement.accessManagementDetails
 
 import app.appContextToInject
 import com.studo.campusqr.common.emailSeparators
